fixed input adaptors for harm and texada

+++ miners/adaptors/output_adaptors/spotltl_to_spotltl/run.sh
@@ -6,5 +6,10 @@ if [ -z "$in_path" ] || [ -z "$out_path" ]; then
     exit 1
 fi
 
+#check that in_path is a file
+if [ ! -f $in_path ]; then
+    echo "Error: $in_path is not a file"
+    exit 1
+fi
 
 cp -r $in_path $out_path
